see extended description :)
- texture refresh sia quando si cambia il file che quando si cambia
l'input;
- possibilità di scegliere "no media" nel load file, con immagine/video
segnaposto per aiutare l'utente a orientarsi tra i formati input/output
- adesso parte senza nulla caricato
- corretto setVolume, che ha cambiato definizione dalla da oF 0072+

+++ src/addons/ELEMENTMAP/ELEMENT_MIXER/elementMixer.h
@@ -61,7 +61,7 @@ class elementMixer : public element
 	int					numOfElements;
 	int*				elementsOrder;
 
-    int                 actualVolume;
+    float                 actualVolume;
 
     int					outputMode;
